Move goth-related fixtures to tests/goth/conftest.py
