+++ tests/goth/conftest.py
@@ -1,0 +1,19 @@
+from datetime import datetime, timezone
+from pathlib import Path
+
+import pytest
+
+
+@pytest.fixture(scope="session")
+def project_dir() -> Path:
+    package_dir = Path(__file__).parent.parent
+    return package_dir.parent.resolve()
+
+
+@pytest.fixture()
+def log_dir() -> Path:
+    base_dir = Path("/", "tmp", "goth-tests")
+    date_str = datetime.now(tz=timezone.utc).strftime("%Y%m%d_%H%M%S%z")
+    log_dir = base_dir / f"goth_{date_str}"
+    log_dir.mkdir(parents=True)
+    return log_dir
